pyproject.toml
[project]
name = "ankify"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai==1.35.0",
    "playwright==1.44.0",
    "python-dateutil==2.8.2",
    "pyyaml==6.0.1",
    "requests==2.31.0",
    "rich==13.7.0",
]
